Split Sparkplug code into files
