+++ acs-monitor/bin/central-monitor.js
@@ -8,7 +8,7 @@ import { ServiceClient } from "@amrc-factoryplus/service-client";
 
 import { GIT_VERSION } from "../lib/git-version.js";
 import { CentralMonitor } from "../lib/central.js";
-import { SparkplugNode } from "../lib/sparkplug.js";
+import { SparkplugNode } from "../lib/sparkplug/node.js";
 
 console.log("Starting ACS central monitor, revision %s", GIT_VERSION);
