+++ acs-monitor/bin/edge-monitor.js
@@ -8,7 +8,7 @@ import { ServiceClient } from "@amrc-factoryplus/service-client";
 
 import { GIT_VERSION } from "../lib/git-version.js";
 import { EdgeMonitor } from "../lib/edge.js";
-import { SparkplugNode } from "../lib/sparkplug.js";
+import { SparkplugNode } from "../lib/sparkplug/node.js";
 
 console.log("Starting ACS edge monitor, revision %s", GIT_VERSION);
